Created Player Class
Created a player class to store player stats.

I made variables private to force control of the variables to the player class.

I did this so whenever a change is made to a variable, it doesn't need to be changed everywhere else. This is also to restrict modifications to variables that should not be modified.